Refactor `algorithms.hpp` (#1354)
* Refactor the functions from `algorithms.hpp`:
  - `mean`: moved to `profile/meter_manager.cpp`
  - `make_index`: moved to `util/index.hpp`
  - `is_strictly_monotonic_increasing`: moved to `/backends/gpu/forest.hpp`
  - `has_contiguous_compartments`:  moved to `/backends/gpu/forest.hpp`.
  - `branches`: moved to `/backends/gpu/forest.hpp`.
  - `expand_branches`: moved to `/backends/gpu/forest.hpp`.
  - `tree_reduce`: moved to `/backends/gpu/forest.hpp`.
  - `is_minimal_degree`: moved to `tree.hpp`.
  - `child_count`: moved to `tree.hpp`.
  - `in_interval`: not used, deleted.
  - `is_strictly_monotonic_decreasing`: not used, deleted.
  - `all_positive`:  not used, deleted.
  - `all_negative`: not used, deleted.
  - `find_branch`: not used, deleted.
  - `is_unique`: not used, deleted.
  - `binary_find`: not used, deleted.
* Remove `make_parent_index` from `tree.hpp`: not used. 
* Replace `make_index` usages with `make_partition`.
* Add unit tests to `test_forest.cpp`, `test_tree.cpp` and `test_index.cpp`.
