+++ arbor/backends/multicore/multi_event_stream.hpp
@@ -13,7 +13,6 @@
 
 #include "backends/event.hpp"
 #include "backends/multi_event_stream_state.hpp"
-#include "algorithms.hpp"
 #include "util/range.hpp"
 #include "util/rangeutil.hpp"
 #include "util/strprintf.hpp"
