+++ test/unit/test_matrix_gpu.cpp
@@ -12,7 +12,6 @@
 
 #include <arbor/math.hpp>
 
-#include "algorithms.hpp"
 #include "matrix.hpp"
 #include "memory/memory.hpp"
 #include "util/span.hpp"
